bug fixed on install area.

+++ AppStore/HTTPPerform.cpp
@@ -1,16 +1,14 @@
 #include "HTTPPerform.h"
-
-
-//const string DOWNLOAD_PATH = "/home/burakmert/Projects/MMIS/DownloaderApp/";
-const string DOWNLOAD_PATH = "/tmp/";
-
-//const string INSTALL_PATH = "/home/burakmert/Projects/MMIS/DownloaderApp/tmpInstall";
-//const string MANIFEST_PATH = "/home/burakmert/Projects/MMIS/DownloaderApp/tmpInstall";
-//const string INSTALL_PATH = "/Downloads/tmpInstall";
-const string INSTALL_PATH = "/usr/bin/";
-
+const string DOWNLOAD_PATH = "/root/AppStore/tmp/Downloads/"; // temporary download path
+const string INSTALL_PATH = "/root/AppStore/tmp/Install/"; // temporary install path
 const string MANIFEST_PATH = "/etc/appmand/";
-
+const string MOVE_PATH = "/usr/bin/"; // binary path final destination
+void clearDirectories()
+{
+    string rmcommand = "rm -rf ";
+    system((rmcommand + INSTALL_PATH).c_str()); // clear tmpInstall
+    system((rmcommand + DOWNLOAD_PATH).c_str()); //clear tmpDownload
+}
 
 void query_updateapps() {
     DBusMessage* msg;
@@ -88,7 +86,7 @@ size_t function_pt(void *contents, size_t size, size_t nmemb, void *stream){
 
 int createManifestFile(application* application)
 {
-    string manifest_dir = MANIFEST_PATH + "/" + application->name + ".mf";
+    string manifest_dir = MANIFEST_PATH + application->name + ".mf";
     cJSON *root;
     root = cJSON_CreateObject();
     cJSON_AddItemToObject(root, "id", cJSON_CreateNumber(application->id + 100));
@@ -117,9 +115,11 @@ int createManifestFile(application* application)
 int calculateHash(application* app)
 {
     FILE* fp;
+    cout << "Opening file :" << app->binaryPath<<endl;
     fp = fopen(app->binaryPath.c_str(),"rb");
     if(fp == NULL)
     {
+        cout << "Fp is null?" << endl;
         app->error = 1;
         app->errorCode = "Could not read binary file";
         return -1;
@@ -152,6 +152,7 @@ int getdir (string dir, application* app)
     DIR *dp;
     struct dirent *dirp;
     string ending = ".sign";
+    string manifest_ending= ".mf";
     if((dp  = opendir(dir.c_str())) == NULL) {
         cout << "Error(" << errno << ") opening " << dir << endl;
         return errno;
@@ -161,7 +162,8 @@ int getdir (string dir, application* app)
         if (strcmp(dirp->d_name,".") == 0  || strcmp(dirp -> d_name,"..") == 0 )
             continue;
         if (dirp->d_type == DT_DIR){
-            string appPath = dir + dirp->d_name;
+            string appPath = INSTALL_PATH + "/" + dirp->d_name;
+            cout << "Calling getdir with : " << appPath << endl;
             getdir(appPath,app);
         }
         else if (string(dirp->d_name).compare(string(dirp->d_name).length() - ending.length(), ending.length(), ending) == 0)
@@ -169,8 +171,13 @@ int getdir (string dir, application* app)
                 if( remove( (dir+"/"+dirp->d_name).c_str() ) != 0 )
                     perror( "Error deleting .sign file" );
             }
+            else if (string(dirp->d_name).compare(string(dirp->d_name).length() - manifest_ending.length(), manifest_ending.length(), manifest_ending) == 0)
+            {
+                continue;  // manifest file, continue
+            }
         else{        /*Set binary path and binary name*/
-            app->binaryPath = dir + "/" +dirp->d_name;
+            app->binaryPath = dir + dirp->d_name;
+            cout << "Binary Path is :" << app->binaryPath << endl;
             app->binaryName =   dirp->d_name;
         }
     }
@@ -232,36 +239,44 @@ int HTTPPerform::download(const string& url, application* app){
     app->error = 1;
     app->isDownloaded = 0;
     app->isInstalled = 0;
-    if (curl){
-        curl_easy_setopt(curl, CURLOPT_URL, url.c_str());
-        curl_easy_setopt(curl, CURLOPT_FOLLOWLOCATION, 1L);
-        string filePath = DOWNLOAD_PATH +  "tmp.tar.gz";
-        fp = fopen(filePath.c_str(),"w+");
-        curl_easy_setopt(curl, CURLOPT_WRITEFUNCTION, write_data);
-        curl_easy_setopt(curl, CURLOPT_WRITEDATA, fp);
-        try {
-            res = curl_easy_perform(curl);
-        }
-        catch(exception &e){
-            cout << "Exception " << e.what() << endl;
-        }
-
-        if(res ==CURLE_OK)
-        {
-            curl_easy_cleanup(curl);
-            fclose(fp);
-            installed = install(filePath, app);
-            if(installed != 0){
-                app->errorCode = "Installation failed\n";
+    string filePath = DOWNLOAD_PATH +  "tmp.tar.gz";
+    fp = fopen(filePath.c_str(),"w+");
+    if (fp == NULL)
+    {
+        app->error;
+        app->errorCode = "Download path does not exist";
+    }
+    else
+    {
+        if (curl){
+            curl_easy_setopt(curl, CURLOPT_URL, url.c_str());
+            curl_easy_setopt(curl, CURLOPT_FOLLOWLOCATION, 1L);
+            curl_easy_setopt(curl, CURLOPT_WRITEFUNCTION, write_data);
+            curl_easy_setopt(curl, CURLOPT_WRITEDATA, fp);
+            try {
+                res = curl_easy_perform(curl);
             }
-            else { // Installation Completed
-                app-> isInstalled = 1;
-                app-> isDownloaded = 1;
-                app-> error = 0;
-                query_updateapps();
-                string command = "rm -r " +filePath;
-                system(command.c_str());
-                returnFlag = 1;
+            catch(exception &e){
+                cout << "Exception " << e.what() << endl;
+            }
+
+            if(res ==CURLE_OK)
+            {
+                curl_easy_cleanup(curl);
+                fclose(fp);
+                installed = install(filePath, app);
+                if(installed != 0){
+                    app->errorCode = "Installation failed\n";
+                }
+                else { // Installation Completed
+                    app-> isInstalled = 1;
+                    app-> isDownloaded = 1;
+                    app-> error = 0;
+                    query_updateapps();
+                    string command = "rm -r " +filePath;
+                    system(command.c_str());
+                    returnFlag = 1;
+                }
             }
         }
     }
@@ -271,13 +286,27 @@ int HTTPPerform::download(const string& url, application* app){
 int HTTPPerform::install(const string& filePath, application* app){
     string command = "tar -zxf " + filePath +" -C  " + INSTALL_PATH + " --strip 1";
     int returnFlag = 0;
-    string dirPath = DOWNLOAD_PATH;
+    string dirPath = INSTALL_PATH;
     int retVal =  system(command.c_str());
     if (retVal!= 0)
+    {
         returnFlag = -1;
+    }
     int dirVal = getdir(dirPath,app);
     if (dirVal != 0 )
+    {
         returnFlag = -1;
+    }
+    string moveCommand = "mv " +app->binaryPath+ " " + MOVE_PATH;
+    cout << "Executing :" << moveCommand << endl;
+    if(system(moveCommand.c_str())!=0)
+        returnFlag = -1; // error while moving application binary to /usr/bin
+    else
+    {
+        app->binaryPath=MOVE_PATH+app->binaryName;
+        cout << "Binary Path changed to :" << app->binaryPath<<endl;
+    }
+
     return returnFlag;
     }
 
@@ -307,7 +336,11 @@ applications* HTTPPerform::perform(ACTION action, int appId){
                     download_url = this->baseUrl + "application/"+to_string(appId)+"/download";
                     installationStatus = this->download(download_url, appList->apps);
                     if(installationStatus != 1)
+                    {
+                        cout << "Installation status is not 1 "<< endl;
+                        clearDirectories();
                         break;
+                    }
                     hashCalculated = calculateHash(appList->apps);
                     if(hashCalculated == -1) {
                         this->errorFlag = 1;
